% Auto-generated: do not edit by hand
\name{interactive2dGraphSmooth}

\alias{interactive2dGraphSmooth}

\title{interactive_2d_graph_smooth component}

\description{

}

\usage{
interactive2dGraphSmooth(id=NULL, additionalData=NULL, additionalDataColor=NULL,
data=NULL, height=NULL, mainDataColor=NULL,
smoothingFactor=NULL, smoothingType=NULL, width=NULL,
xLabel=NULL, yLabel=NULL)
}

\arguments{
\item{id}{Character. }

\item{additionalData}{List of lists containing elements 'x', 'y'.
those elements have the following types:
  - x (numeric; optional)
  - y (numeric; optional)s. }

\item{additionalDataColor}{List of characters. }

\item{data}{List of lists containing elements 'x', 'y'.
those elements have the following types:
  - x (numeric; required)
  - y (numeric; required)s. }

\item{height}{Numeric. }

\item{mainDataColor}{Character. }

\item{smoothingFactor}{Numeric. }

\item{smoothingType}{Character. }

\item{width}{Numeric. }

\item{xLabel}{Character. }

\item{yLabel}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

