% Auto-generated: do not edit by hand
\name{interactive2dGraphSmoth}

\alias{interactive2dGraphSmoth}

\title{interactive_2d_graph_smoth component}

\description{

}

\usage{
interactive2dGraphSmoth(id=NULL, height=NULL, points=NULL, width=NULL, xLabel=NULL,
yLabel=NULL)
}

\arguments{
\item{id}{Character. }

\item{height}{Numeric. }

\item{points}{List of lists containing elements 'x', 'y'.
those elements have the following types:
  - x (numeric; required)
  - y (numeric; required)s. }

\item{width}{Numeric. }

\item{xLabel}{Character. }

\item{yLabel}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

